pyproject.toml
[project]
name = "mmaction2-streamlit"
version = "0.1.0"
description = "Streamlit application for action recognition with MMAction2"
authors = [{ name = "Contributors" }]
requires-python = ">=3.10,<3.12"
readme = "README.md"
dependencies = [
    "streamlit>=1.30",
    "torch>=2.1",
    "torchvision>=0.16",
    "mmaction2>=1.1",
    "mmengine>=0.7",
    "mmcv>=2.1,<2.2",
    "opencv-python>=4.7",
    "numpy>=1.24",
    "pydantic>=1.10",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "requests>=2.31",
    "openmim>=0.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "black>=24.3"
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app", "core", "models"]
exclude = ["tests", "tests.*", "assets", "configs", "scripts"]

[tool.setuptools.package-data]
app = ["*.py"]
core = ["*.py"]
models = ["*.py"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-ra"

[tool.black]
target-version = ["py310"]
line-length = 100

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "I", "PL", "UP", "B"]
ignore = ["E501"]
